feat(bake/oracle): Added oracle OCI bakery.
